Make it compile under other OS's (hopefully) also add some minor tweaks
AppleOnlyMethods has been renamed VersionDependentMethodUtilities to
reflect what it does now
It can now be called directly, because a library (AppleJavaExtensions)
has been added as an empty implementation of the com.apple.eawt stff
This can be gotten from
<https://developer.apple.com/legacy/library/samplecode/AppleJavaExtensions/Introduction/Intro.html#//apple_ref/doc/uid/DTS10000677-Intro-DontLinkElementID_2>,
but should be included (in the lib folder) and should automatically set
up with maven.  
The version number in the pom for AppleJavaExtensions is not correct,
and needs to be updated when I have an internet connection.